auto_generate_video_fold6/services/graphql-gateway/requirements.txt
# FastAPI 核心套件
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# GraphQL 支援
strawberry-graphql[fastapi]>=0.213.0
graphql-core>=3.2.0

# 資料庫連接
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# 身份驗證
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP 客戶端
httpx>=0.25.0
aiohttp>=3.9.0

# 日誌記錄
structlog>=23.2.0

# 效能監控
prometheus-client>=0.19.0

# 開發工具
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0